// src/app/models/surveyResult.ts
export class SurveyResult {
    block1: string;
    block1Value: number;
    block1Percentage: number;
    block1Name: string;
    block2: string;
    block2Value: number;
    block2Percentage: number;
    block2Name: string;
    block3: string;
    block3Value: number;
    block3Percentage: number;
    block3Name: string;
    global: string;
    globalValue: number;
    globalPercentage: number;
    datetime: string;
}